DPdashboard-main1/DPdashboard-main/requirements.txt
beautifulsoup4==4.11.1
Brotli==1.0.9
bs4==0.0.1
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
dash==2.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
facebook-sdk==3.1
Flask==2.1.3
Flask-Compress==1.12
google-api-core==2.8.2
google-api-python-client==2.54.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
httplib2==0.20.4
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.1
oauth2client==4.1.3
oauthlib==3.2.0
pandas==1.4.3
plotly==5.9.0
protobuf==4.21.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.3.2.post1
tenacity==8.0.1
tqdm==4.64.0
uritemplate==4.1.1
urllib3==1.26.10
Werkzeug==2.1.2
